projectmain/requirements.txt
Django
django-allauth
djangorestframework
Pillow
